Added Front and Backend Push functions
client.js and expoPushToken.js on the frontend are sending the
Expo auth servers the token we generate in TabNavigator.js.

TabNavigator is handling generation of push tokens for the device
every time someone enters the logged-in stack - does not overwrite,
stays the same per device.

On the backend, pushNotifications sends to the Expo auth servers,
voteHandler is generating the payload for it, commonsVotes is fetching
for voteHandler from commons API.